feat(config): add model field to ReviewConfig and DebateConfig
Review and debate workflows were not passing --model to gemini-cli,
causing model steering to select Flash instead of Pro. Root cause:
apply_tool_defaults=false (by design) blocks [tools.gemini-cli].default_model,
and ReviewConfig/DebateConfig had no model field of their own.

Changes:
- Add model: Option<String> to ReviewConfig and DebateConfig
- Add Executor::override_model() for explicit model > tier model_spec
- Wire model resolution in review_cmd.rs and debate_cmd.rs
  (CLI --model > config model > tier model_spec > tool default)
- Update build_executor to apply model override when model_spec present
- Split global_tests.rs to stay under 800-line limit
- Add 9 new tests (config parsing, is_default, serialization, override)
