+++ Cargo.toml
@@ -3,7 +3,7 @@ members = ["crates/*"]
 resolver = "2"
 
 [workspace.package]
-version = "0.1.107"
+version = "0.1.108"
 edition = "2024"
 rust-version = "1.85"
 license = "Apache-2.0"
